+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 **Changelog** (v0.00.10-beta 05-05-2024):
 - Fixed Get Ip Modem Rakitan
+- Add Nama Modem For Log Telegram
+- Fix Typo
 
 **Changelog** (v0.00.09-beta 04-05-2024):
 - Fixed File Bot Not Found